requirements.txt
blinker==1.9.0
cachelib==0.13.0
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
chess==1.11.2
click==8.2.1
cssselect2==0.8.0
defusedxml==0.7.1
Flask==3.1.2
Flask-Caching==2.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pycparser==2.22
python-chess==1.999
python-dotenv==1.1.1
stockfish==3.28.0
tinycss2==1.4.0
webencodings==0.5.1
Werkzeug==3.1.3
